+++ Changelog.md
@@ -1,3 +1,8 @@
+
+# Master
+
+- Enables to use --show-software and --show-path together in [*sams-software-updater*](docs/sams-software-updater.md).
+
 # Version 1.2 - 2020-04-06
 
 - Set umask in deamon mode. See: [*sams-collector*](docs/sams-collector.md)